@@ -1,11 +1,11 @@
 <Project Sdk="Microsoft.NET.Sdk">
 	<PropertyGroup>
-	  <TargetFrameworks>netstandard2.1;net6.0;net8.0</TargetFrameworks>
+	  <TargetFrameworks>netstandard2.1;net8.0</TargetFrameworks>
 	  <LangVersion>12.0</LangVersion>
 	  <AssemblyName>DarkPeakLabs.RdapClient</AssemblyName>
 
 	  <PackageId>DarkPeakLabs.RdapClient</PackageId>
-	  <Version>1.0.2</Version>
+	  <Version>1.1.0</Version>
 	  <Authors>DarkPeakLabs</Authors>
 	  <Company>DarkPeakLabs</Company>
 	  <Product>DarkPeakLabs.RdapClient</Product>
@@ -32,10 +32,6 @@
     <PackageReference Include="System.ComponentModel.Annotations" />
     <PackageReference Include="System.Text.Json" />
     <PackageReference Include="LazyCache" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="Client\" />
   </ItemGroup>
 
   <ItemGroup>
