@@ -1,8 +1,5 @@
-﻿using Microsoft.Extensions.Logging;
-using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Globalization;
-using System.Text.Json;
 
 namespace DarkPeakLabs.Rdap.Conformance
 {
@@ -11,11 +8,6 @@ namespace DarkPeakLabs.Rdap.Conformance
     /// </summary>
     public class RdapConformance
     {
-        /// <summary>
-        /// Class logger
-        /// </summary>
-        private ILogger _logger;
-
         private List<RdapConformanceViolation> violations;
 
         /// <summary>
@@ -31,10 +23,14 @@ namespace DarkPeakLabs.Rdap.Conformance
         /// <summary>
         /// Creates a new RDAP conformance instance
         /// </summary>
-        internal RdapConformance(ILogger logger = null)
+        internal RdapConformance()
         {
-            _logger = logger;
-            violations = new List<RdapConformanceViolation>();
+            violations = [];
+        }
+
+        internal void AddViolation(RdapConformanceViolation violation)
+        {
+            violations.Add(violation);
         }
 
         /// <summary>
@@ -43,24 +39,9 @@ namespace DarkPeakLabs.Rdap.Conformance
         /// <param name="severity">Violation severity</param>
         /// <param name="category">Violation category</param>
         /// <param name="message">Violation message</param>
-        public void AddViolation(RdapConformanceViolationSeverity severity, RdapConformanceViolationCategory category, string message)
+        internal void AddViolation(RdapConformanceViolationSeverity severity, RdapConformanceViolationCategory category, string message)
         {
-            switch (severity)
-            {
-                case RdapConformanceViolationSeverity.Error:
-                    _logger?.LogError("Conformance violation. Category = {Category}, Message = {Message}", category, message); break;
-                case RdapConformanceViolationSeverity.Information:
-                    _logger?.LogInformation("Conformance violation. Category = {Category}, Message = {Message}", category, message); break;
-                case RdapConformanceViolationSeverity.Warning:
-                    _logger?.LogWarning("Conformance violation. Category = {Category}, Message = {Message}", category, message); break;
-            }
-
-            violations.Add(new RdapConformanceViolation()
-            {
-                Severity = severity,
-                Category = category,
-                Issue = message
-            });
+            AddViolation(new RdapConformanceViolation(severity, category, message));
         }
 
         /// <summary>
@@ -68,30 +49,13 @@ namespace DarkPeakLabs.Rdap.Conformance
         /// </summary>
         /// <param name="severity">Violation severity</param>
         /// <param name="category">Violation category</param>
-        /// <param name="message">Violation message</param>
         /// <param name="format">Message format string</param>
         /// <param name="args">Message arguments</param>
-        internal void AddViolation(RdapConformanceViolationSeverity severity, RdapConformanceViolationCategory category, string format, params object[] args) => AddViolation(severity, category, string.Format(CultureInfo.InvariantCulture, format, args));
-
-        /// <summary>
-        /// Adds new JSON response violation to the list
-        /// </summary>
-        /// <param name="severity">Violation severity</param>
-        /// <param name="reader">JSON reader</param>
-        /// <param name="message">Violation message</param>
-        internal void AddJsonViolation(RdapConformanceViolationSeverity severity, ref Utf8JsonReader reader, string message)
-        {
-            AddViolation(severity, RdapConformanceViolationCategory.JSON, $"JSON start index: {reader.TokenStartIndex}, {message}");
-        }
-
-        /// <summary>
-        /// Adds new JSON response violation to the list
-        /// </summary>
-        /// <param name="severity">Violation severity</param>
-        /// <param name="reader">JSON reader</param>
-        /// <param name="format">Message format string</param>
-        /// <param name="args">Message arguments</param>
-        internal void AddJsonViolation(RdapConformanceViolationSeverity severity, ref Utf8JsonReader reader, string format, params object[] args) => AddJsonViolation(severity, ref reader, string.Format(CultureInfo.InvariantCulture, format, args));
+        internal void AddViolation(
+            RdapConformanceViolationSeverity severity,
+            RdapConformanceViolationCategory category,
+            string format,
+            params object[] args) => AddViolation(severity, category, string.Format(CultureInfo.InvariantCulture, format, args));
 
         /// <summary>
         /// Adds new web service implementation violation to the list
@@ -109,11 +73,9 @@ namespace DarkPeakLabs.Rdap.Conformance
         /// <param name="severity">Violation severity</param>
         /// <param name="format">Message format string</param>
         /// <param name="args">Message arguments</param>
-        internal void AddImplementationViolation(RdapConformanceViolationSeverity severity, string format, params object[] args) => AddImplementationViolation(severity, string.Format(CultureInfo.InvariantCulture, format, args));
-
-        public void AddViolations(IReadOnlyList<RdapConformanceViolation> violations)
-        {
-            this.violations.AddRange(violations);
-        }
+        internal void AddImplementationViolation(
+            RdapConformanceViolationSeverity severity,
+            string format,
+            params object[] args) => AddImplementationViolation(severity, string.Format(CultureInfo.InvariantCulture, format, args));
     }
 }
