@@ -1,4 +1,6 @@
 ﻿using System;
+using System.Net.Http;
+using System.Net.Http.Headers;
 using DarkPeakLabs.Rdap.Conformance;
 
 namespace DarkPeakLabs.Rdap
@@ -8,7 +10,10 @@ namespace DarkPeakLabs.Rdap
         /// <summary>
         /// Raw string Json response from last successful request
         /// </summary>
-        public string RawJson { get; internal set; }
+        public string HttpResponseBody { get; internal set; }
+        public HttpResponseHeaders HttpResponseHeaders { get; internal set; }
+        public HttpRequestHeaders HttpRequestHeaders { get; internal set; }
+        public Version HttpVersion { get; internal set; }
 
         /// <summary>
         ///
@@ -17,10 +22,13 @@ namespace DarkPeakLabs.Rdap
 
         public T Value { get; internal set; }
 
-        public DateTime? RequestSent { get; internal set; }
-        public DateTime? ResponseReceived { get; internal set; }
-        public DateTime? RequestFailed { get; internal set; }
-        public DateTime? DataReadStarted { get; internal set; }
-        public DateTime? DataReadFinished { get; internal set; }
+        public DateTimeOffset? RequestSent { get; internal set; }
+        public DateTimeOffset? ResponseReceived { get; internal set; }
+        public DateTimeOffset? RequestFailed { get; internal set; }
+        public DateTimeOffset? DataReadStarted { get; internal set; }
+        public DateTimeOffset? DataReadFinished { get; internal set; }
+        public HttpMethod HttpRequestMethod { get; internal set; }
+        public Uri HttpRequestUri { get; internal set; }
+        public HttpContentHeaders HttpContentHeaders { get; internal set; }
     }
 }
