@@ -1,4 +1,5 @@
 ﻿using CsvHelper;
+using CsvHelper.Configuration;
 using System;
 using System.Collections.Generic;
 using System.Globalization;
@@ -48,10 +49,17 @@ public class CsvHttpClient : HttpClient
     /// <typeparam name="T">Record data type</typeparam>
     /// <param name="requestUri">Request Url</param>
     /// <returns></returns>
-    public async Task<IReadOnlyCollection<T>> GetRecordsAsync<T>(Uri requestUri)
+    public async Task<IReadOnlyList<T>> GetRecordsAsync<T>(Uri requestUri)
     {
         using StreamReader reader = new StreamReader(await GetStreamAsync(requestUri).ConfigureAwait(false));
-        using CsvReader csvReader = new CsvReader(reader, CultureInfo.InvariantCulture);
+        var csvConfiguration = new CsvConfiguration(CultureInfo.InvariantCulture)
+        {
+            Delimiter = ",",
+            HasHeaderRecord = true,
+            TrimOptions = TrimOptions.Trim,
+            MissingFieldFound = null
+        };
+        using CsvReader csvReader = new CsvReader(reader, csvConfiguration);
         return csvReader.GetRecords<T>().ToList();
     }
 
@@ -61,7 +69,7 @@ public class CsvHttpClient : HttpClient
     /// <typeparam name="T">Record data type</typeparam>
     /// <param name="requestUri">Request Url</param>
     /// <returns></returns>
-    public IReadOnlyCollection<T> GetRecords<T>(Uri requestUri)
+    public IReadOnlyList<T> GetRecords<T>(Uri requestUri)
     {
         return GetRecordsAsync<T>(requestUri).GetAwaiter().GetResult();
     }
