@@ -46,16 +46,21 @@ internal sealed class Program
     private static void GenerateCode(string solutionRoot)
     {
         var code = RdapEnumCodeGenerator.GenerateLinkRelationEnumAsync().GetAwaiter().GetResult();
-        code.SaveAs(Path.Combine(solutionRoot, @"Rdap\Values\Json\RdapLinkRelationType.cs"));
+        code.SaveAs(Path.Combine(solutionRoot, @"package\RdapClient\Values\RdapLinkRelationType.cs"));
 
         code = RdapEnumCodeGenerator.GenerateStatusEnumAsync().GetAwaiter().GetResult();
-        code.SaveAs(Path.Combine(solutionRoot, @"Rdap\Values\Json\RdapStatus.cs"));
+        code.SaveAs(Path.Combine(solutionRoot, @"package\RdapClient\Values\RdapStatus.cs"));
 
         code = RdapEnumCodeGenerator.GenerateEventActionEnumAsync().GetAwaiter().GetResult();
-        code.SaveAs(Path.Combine(solutionRoot, @"Rdap\Values\Json\RdapEventAction.cs"));
+        code.SaveAs(Path.Combine(solutionRoot, @"package\RdapClient\Values\RdapEventAction.cs"));
         code = RdapEnumCodeGenerator.GenerateEntityRoleEnumAsync().GetAwaiter().GetResult();
-        code.SaveAs(Path.Combine(solutionRoot, @"Rdap\Values\Json\RdapEntityRole.cs"));
+        code.SaveAs(Path.Combine(solutionRoot, @"package\RdapClient\Values\RdapEntityRole.cs"));
         code = RdapEnumCodeGenerator.GenerateNoticeAndRemarkTypeEnumAsync().GetAwaiter().GetResult();
-        code.SaveAs(Path.Combine(solutionRoot, @"Rdap\Values\Json\RdapNoticeAndRemarkType.cs"));
+        code.SaveAs(Path.Combine(solutionRoot, @"package\RdapClient\Values\RdapNoticeAndRemarkType.cs"));
+
+        code = RdapEnumCodeGenerator.GenerateDnsSecAlgorithmNumbersAsync().GetAwaiter().GetResult();
+        code.SaveAs(Path.Combine(solutionRoot, @"package\RdapClient\Values\DnsSecAlgorithmType.cs"));
+        code = RdapEnumCodeGenerator.GenerateDnsSecDigestTypesAsync().GetAwaiter().GetResult();
+        code.SaveAs(Path.Combine(solutionRoot, @"package\RdapClient\Values\DnsSecDigestType.cs"));
     }
 }
